renamed the `hardware` module -> `architectures`
