+++ kernel/src/library/architectures/_x86_64/boot/linker.ld
